BUILDENG-5324 Use rest_client instead of net/http
When using net/http, every request was rejected with a 404 Not Found.
I never found the actual problem, but I suspect that the way the
credentials are passed along (encoded as http://user:pass@...) doesn't
work well with Nexus' REST api. Changing to rest_client and using
'explicit' authentication seem to solve the problem. Encoding the
credentials in the url failed with the rest_client as well.